+Xong chức năng tạo hóa đơn Đã xác nhận

+++ app/src/main/java/com/dinhconghien/getitapp/Models/ListLaptop.kt
@@ -4,7 +4,7 @@ data class ListLaptop(
     val idLap : String ="",
     var idBrandLap : String = "",
     var nameLap : String = "",
-    var priceLap : Int = 0,
+    var priceLap : Long = 0,
     var quantity : Int =0,
     var avaLap : String ="",
     var rating : Int = 0,
